requirements.txt
streamlit
pandas
openpyxl
python-docx
num2words
